+++ src/ppidd_ga_mpi.cpp
@@ -340,7 +340,7 @@ static int n_in_msg_mpiq=0;
    }
 
 
-   int PPIDD_Location(int handle, int64_t *ilo, int64_t *ihi, int64_t *map, int64_t *proclist, int *np) {
+   int PPIDD_Location(int handle, int64_t *ilo, int64_t *ihi, int64_t *map, int *proclist, int *np) {
       ga_int mpiilo[1]={(ga_int)*ilo-1};
       ga_int mpiihi[1]={(ga_int)*ihi-1};
 
@@ -353,7 +353,7 @@ static int n_in_msg_mpiq=0;
       for (int i=0;i<*np;i++) {
          map[2*i]=(int64_t)(mpimap[2*i]+1);
          map[2*i+1]=(int64_t)(mpimap[2*i+1]+1);
-         proclist[i]=(int64_t)mpiproclist[i];
+         proclist[i]=mpiproclist[i];
       }
       return 1 ;
    }
